pyproject.toml
[project]
name = "movie-db"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "neo4j==5.15.0",
    "openai>=1.3.7",
    "pydantic>=2.7.2",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "uvicorn>=0.30.0",
    "mcp[cli]>=1.10.0",
]
